migrate current js to ts
